Resolve "use pydantic: model_validate / model_validate_json instead of Model(**object)"

Closes #87

See merge request swiss-armed-forces/cyber-command/cea/loom!50

+++ backend/api/api/models/statistics_model.py
@@ -65,7 +65,7 @@ class SummaryStatisticsModel(BaseModel):
     count: int
     min: int
     max: int
-    avg: float
+    avg: int
 
     @staticmethod
     def from_statistics_summary(stats: StatisticsSummary):
